Rename: "트랜잭션 위치 수정"
트랜잭션 util 위치 변경에 따른 import 코드 수정
