+++ routes/squad.router.js
@@ -1,6 +1,6 @@
 import express from "express";
 import { prisma } from "../utils/prisma/prismaClient.js";
-import { executeTransaction } from "../utils/executeTransaction.js";
+import { executeTransaction } from "../utils/transaction/executeTransaction.js";
 
 const router = express.Router();
